0.9.3
Fix error meta missing when instantiating a thrown error

When instantiating a new error using an existing error as an option, the meta of the former error was
not retained. This was causing issues because the engine would throw a new error if an action failed losing the context of the failed action's error.

Resolves: #29