Implement the destroy method to be more persistent
The original assumption was that the `QUIT` command was in the execution
stack of the blocked client and would pick it up once the special item
is pushed to the queue; It seems that there are still occasion where
the client would hug the thread.

It's possible that the stack has more `BRPOP` than `QUIT`, if that's the
case, this should be investigated and this method be reimplemented.